text field with visual currency formatter

+++ source/app/src/main/java/com/charmflex/flexiexpensesmanager/features/account/data/entities/AccountEntity.kt
@@ -27,7 +27,7 @@ internal data class AccountEntity(
     @ColumnInfo("is_deleted", defaultValue = "false")
     val isDeleted: Boolean = false,
     @ColumnInfo("initial_amount", defaultValue = "0")
-    val initialAmount: Int = 0,
+    val initialAmount: Long = 0,
     @Embedded
     val additionalInfo: AdditionalInfo?
 ) {
