+++ source/app/src/main/java/com/charmflex/flexiexpensesmanager/features/transactions/data/repositories/TransactionRepositoryImpl.kt
@@ -21,7 +21,7 @@ internal class TransactionRepositoryImpl @Inject constructor(
     override suspend fun addNewExpenses(
         name: String,
         fromAccountId: Int,
-        amount: Int,
+        amount: Long,
         categoryId: Int,
         transactionDate: String,
     ) {
@@ -40,7 +40,7 @@ internal class TransactionRepositoryImpl @Inject constructor(
     override suspend fun addNewIncome(
         name: String,
         toAccountId: Int,
-        amount: Int,
+        amount: Long,
         categoryId: Int,
         transactionDate: String
     ) {
@@ -60,7 +60,7 @@ internal class TransactionRepositoryImpl @Inject constructor(
         name: String,
         fromAccountId: Int,
         toAccountId: Int,
-        amount: Int,
+        amount: Long,
         transactionDate: String
     ) {
         val transaction = TransactionEntity(
