+++ source/app/src/main/java/com/charmflex/flexiexpensesmanager/features/transactions/domain/repositories/TransactionRepository.kt
@@ -9,7 +9,7 @@ internal interface TransactionRepository {
     suspend fun addNewExpenses(
         name: String,
         fromAccountId: Int,
-        amount: Int,
+        amount: Long,
         categoryId: Int,
         transactionDate: String,
     )
@@ -17,7 +17,7 @@ internal interface TransactionRepository {
     suspend fun addNewIncome(
         name: String,
         toAccountId: Int,
-        amount: Int,
+        amount: Long,
         categoryId: Int,
         transactionDate: String,
     )
@@ -26,7 +26,7 @@ internal interface TransactionRepository {
         name: String,
         fromAccountId: Int,
         toAccountId: Int,
-        amount: Int,
+        amount: Long,
         transactionDate: String,
     )
     suspend fun getHistory(): List<ExpensesData>
